Port `load_zmq` and `save_zmq`
